+++ cas-server/src/main/java/com/maoatao/cas/core/bean/param/clientscopepermission/ClientScopePermissionQueryParam.java
@@ -40,7 +40,7 @@ public class ClientScopePermissionQueryParam extends BasePaginationParam {
      */
     @NotNull(message = "scopeId 不能为空")
     @Schema(description = "用户id")
-    private Integer scopeId;
+    private Long scopeId;
     /**
      * 权限id
      */
