+++ cas-server/src/main/java/com/maoatao/cas/core/bean/param/clientscopepermission/ClientScopePermissionSaveParam.java
@@ -34,7 +34,7 @@ public class ClientScopePermissionSaveParam extends BaseSaveParam {
      */
     @NotNull(message = "scopeId 不能为空")
     @Schema(description = "用户id")
-    private Integer scopeId;
+    private Long scopeId;
     /**
      * 权限id
      */
